Add a hack to throttle orientation sensor input events
The userspace orientation sensor code in Android ICS seems to constantly
send updates through the uinput subsystem, which causes the CPU to
constantly run at 15-20% while the system is awake.  This is almost
certainly a userspace bug, but since the orientation sensor driver is
proprietary (ugh), work around this by forcing the driver to sleep after
each update.

Signed-off-by: Steven Luo <steven+android@steven676.net>